+++ src/HSA/App.xaml.cs
@@ -114,6 +114,8 @@ public partial class App : Application
         sc.AddSingleton<IPrinterService, PrinterService>();
         sc.AddSingleton<IDriverService, DriverService>();
         sc.AddSingleton<IFirmwareService, FirmwareService>();
+        sc.AddSingleton<IppConsumableSource>();
+        sc.AddSingleton<PrinterEndpointDiscovery>();
         sc.AddSingleton<IConsumableService, ConsumableService>();
         sc.AddSingleton<IModelImageService, ModelImageService>();